Move some of the method in DateTimeInfoParser to DateTimeInfo
